requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dateutil==2.9.0.post0
pyswisseph==2.10.3.2
mangum

